[MOB-3897] Move video player to separate file
